fix `listen EADDRINUSE: address already in use :::8080` when api tests are running on CI, with coverage cf https://github.com/openwhyd/openwhyd/actions/runs/6000177975/job/16271736742#step:6:266

+++ test/approval-tests-helpers.js
@@ -158,14 +158,10 @@ const startOpenwhydServerWith = async (env) =>
         if (serverProcess.killed) return resolve();
         serverProcess.on('close', resolve);
         if (serverProcess.kill(/*'SIGTERM'*/)) return; // we will resolve when the `close` event is received
-        console.warn('🧟‍♀️ failed to kill childprocess!');
-        if (serverProcess.pid) {
-          try {
-            process.kill(-serverProcess.pid, 'SIGINT');
-          } catch (err) {
-            console.warn('failed to kill by pid:', err.message);
-          }
+        if (!serverProcess.killed && serverProcess.pid) {
+          process.kill(-serverProcess.pid, 'SIGINT');
         }
+        throw new Error('🧟‍♀️ failed to kill childprocess!');
       });
     serverProcess.on('error', reject);
     serverProcess.stderr.on('data', errPrinter);
